+++ fv3core/fv3core/stencils/a2b_ord4.py
@@ -13,10 +13,10 @@ from gt4py.gtscript import (
 
 import pace.dsl.gt4py_utils as utils
 from fv3core.stencils.basic_operations import copy_defn
-from pace.dsl.stencil import StencilFactory
+from pace.dsl.stencil import GridIndexing, StencilFactory
 from pace.dsl.typing import FloatField, FloatFieldI, FloatFieldIJ
-from pace.stencils.testing.grid import GridData, GridIndexing, axis_offsets
 from pace.util import X_DIM, X_INTERFACE_DIM, Y_DIM, Y_INTERFACE_DIM, Z_DIM
+from pace.util.grid import GridData
 
 
 # comact 4-pt cubic interpolation
@@ -574,8 +574,7 @@ class AGrid2BGridFourthOrder:
         )
         origin, domain = self._exclude_tile_edges(origin, domain)
 
-        ax_offsets = axis_offsets(
-            self._idx,
+        ax_offsets = self._idx.axis_offsets(
             origin,
             domain,
         )
