+++ fv3gfs-physics/fv3gfs/physics/stencils/microphysics.py
@@ -9,7 +9,7 @@ import pace.dsl.gt4py_utils as utils
 import pace.util.constants as constants
 from pace.dsl.stencil import StencilFactory
 from pace.dsl.typing import Float, FloatField, FloatFieldIJ, Int
-from pace.stencils.testing.grid import GridData
+from pace.util.grid import GridData
 
 
 def fields_init(
